Refactor/#20: pr 수정사항 반영
