+++ cluvr-api/src/main/java/com/example/cluvrapi/domain/analytics/entity/CategoryStat.java
@@ -7,11 +7,14 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 
+import lombok.AccessLevel;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Entity
 @Getter
 @Table(name = "category_stat")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class CategoryStat {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -31,5 +34,15 @@ public class CategoryStat {
 
 	@Column(nullable = false)
 	private Long categoryId;
+
+	public CategoryStat(Integer totalAnswer, Integer totalSelected, Integer totalScore,
+		Integer totalQuestion, Long userId, Long categoryId) {
+		this.totalAnswer = totalAnswer;
+		this.totalSelected = totalSelected;
+		this.totalScore = totalScore;
+		this.totalQuestion = totalQuestion;
+		this.userId = userId;
+		this.categoryId = categoryId;
+	}
 
 }
