+++ cluvr-api/src/main/java/com/example/cluvrapi/domain/clover/dto/request/CreateCloverRequestDto.java
@@ -2,7 +2,7 @@ package com.example.cluvrapi.domain.clover.dto.request;
 
 import lombok.Getter;
 
-import com.example.cluvrapi.domain.clover.eunms.Tier;
+import com.example.cluvrapi.domain.clover.enums.Tier;
 
 @Getter
 public class CreateCloverRequestDto {
