+++ cluvr-api/src/main/java/com/example/cluvrapi/domain/clover/enums/Tier.java
@@ -1,0 +1,5 @@
+package com.example.cluvrapi.domain.clover.enums;
+
+public enum Tier {
+	CLOVER1, CLOVER2, CLOVER3
+}
