--- cluvr-api/src/main/java/com/example/cluvrapi/domain/clover/eunms/Tier.java
@@ -1,5 +1,0 @@
-package com.example.cluvrapi.domain.clover.eunms;
-
-public enum Tier {
-	CLOVER1, CLOVER2, CLOVER3
-}
